pyproject.toml
[project]
name = "obs-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "obsws-python>=1.7.1",
    "python-dotenv>=1.1.0",
]
